% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADNI_unzip.R
\name{ADNI_unzip}
\alias{ADNI_unzip}
\title{Wrapper function to unzip ADNI dataset}
\usage{
ADNI_unzip(dest_folder = "data_unziped")
}
\arguments{
\item{dest_folder}{Name of destination folder for decompressed/unziped files.}
}
\value{
Side effect, files are decompressed and placed in persistent disk.
}
\description{
Wrapper function to unzip ADNI dataset
}
