% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insync_data.R
\name{insync_data}
\alias{insync_data}
\title{Synchronizing google bucket contents into Terra persistent disk.}
\usage{
insync_data(source_bucket, folder_name = "data")
}
\arguments{
\item{source_bucket}{A string representing the google bucket id.}

\item{folder_name}{The folder name as it appears in the Terra Workspace.}
}
\value{
Side effect; synchronized folder contents.
}
\description{
Synchronizing google bucket contents into Terra persistent disk.
}
