% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_contents.R
\name{unzip_contents}
\alias{unzip_contents}
\title{Unzip content of a file or folder containing zipped files}
\usage{
unzip_contents(file = NULL)
}
\arguments{
\item{file}{A link to Google bucket folder or file to be unzipped}
}
\value{
A folder with contents
}
\description{
Unzip content of a file or folder containing zipped files
}
